[FTR] - Continue the challenge: - Code refactored.
